Paint point selections using CTRL+LMB click/drag, WIP
This selection method makes it relatively easy to select contiguous
regions of point geometry.  It's quite like painting on the point cloud.

Currently a work in progress, since this doesn't scale to large point
clouds due to the way that cursor snapping is implemented.  Also there's
some additional cleanup required.